Typo fix
I would have suggested "out" to be "our" as the most readily available fix, but open source ownership trickier than that. You could state it as "our" and sound closed to contributes, "your" and make the project sound like an abandoned code dump, or "the result of core team and community contributions", which is both verbose and makes an "us vs them" comparison for contributors, etc. Thus I chose to describe the model as a component of the project rather than a component owned by some group of people.