A major refactoring of the input system.
Yes, this still uses lots of flags, but has decoupled mostly the
input system from the model. (There are still a few hard coded
'comment_text' feature bindings in the keras directories.)

The decision to hard code the 'text' feature in the simple case
as is used in tf_hub is a reflection of the way our analyzers are
currently coded.

There are several fixes to things that are currently broken,
and all of these models have been tested locally.