Address comments that resulted from meeting with Lucas.
This changes the model we build to always create a 'text' and/or 'tokens'
feature for input to the evaluation. This is uniform and does not
vary with training set - which is more predictable.
These are done via hard coded constants in the base_model file.

The training-set-specific parameters are now all concentrated in
tfrecord_simple, and will eventually be made into a class that can
handle fetching and downloading or whatever for particular cases.

All models were test up to successful training.